memory_opt_gadgets: move code into constrain_word_rlc()
